Fixed CRT Shader for Windows
The problem was the amount of varying variable used in that shader.
Since in Windows, WebGL goes through ANGLE, it is likely that it cannot
support as many variables. Anyway, it is all good since a lot of these
variables should have been made exposed as uniforms in the first place.