Improved Entity2D's animation and resize
The JSON specification for sprite's animation has changed allowing to provide a duration in seconds for each frame. There was a few little quirks in the way the resizing was handled when the game resolution changed. It is is still more complex than I wish but it all works well for the user.